engine/backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
anthropic==0.40.0
httpx==0.27.0
redis==5.0.1
pyyaml==6.0.1
pytz==2024.1
werkzeug==3.0.1
python-dotenv==1.0.0
stripe==7.0.0
sentry-sdk[flask]==1.39.1
